+++ packages/frontend/src/features/dashboardTilecomments/components/CommentTimestamp.tsx
@@ -1,0 +1,24 @@
+import { Comment } from '@lightdash/common';
+import { Text, Tooltip } from '@mantine/core';
+import moment from 'moment';
+import { FC } from 'react';
+import { useTimeAgo } from '../../../hooks/useTimeAgo';
+
+type Props = {
+    timestamp: Comment['createdAt'];
+};
+
+export const CommentTimestamp: FC<Props> = ({ timestamp }) => {
+    const timeAgo = useTimeAgo(timestamp);
+
+    return (
+        <Tooltip
+            position="top-start"
+            label={moment(timestamp).format('YYYY-MM-DD HH:mm:ss')}
+        >
+            <Text fz="xs" color="gray.5">
+                {timeAgo}
+            </Text>
+        </Tooltip>
+    );
+};
