+++ packages/frontend/src/features/dashboardTilecomments/utils/getNameInitials.ts
@@ -1,0 +1,5 @@
+export const getNameInitials = (name: string) =>
+    name
+        .split(' ')
+        .map((n) => n[0])
+        .join('');
